Workaround for undetected abnormal server disconnection
For solving the abnormal tcf connection server down, such as closing
qemu rudely, the TCF channel is undetected the server down. With
netstat, we can see the TCF channel is still alive since the TCF
connection timeout is 2 hours! So the workaround here is that we
give the shell service one time retry if meeting connection error.
It will suppose that connection is destroyed and we need to re-
subscribe the stream service. So that we have a chance of success.
It will improve the user experience that even after shutdown qemu
and restart qemu again, we can still debugging without meeting connection
failed error(we need not manually disconnect the TCF connection in RSE
view any more).
And also, improve the error message so that user will not be
messed up by the incorrect error messge.

Signed-off-by: Liping ke<liping.ke@intel.com>
Signed-off-by: Lianhao Lu<lianhao.lu@intel.com>