Yocto-BSP Wizard-time consuming processes handling
- problems on the main page of yocto-bsp wizard :
        - getting the available kernel architectures is invoked from the UI
                -> similar to the properties page, when loading the
                architectures a new thread is created in the background and a
progress monitor dialog is shown.
        - getting qemu kernel architectures is invoked from the UI
                -> as above.
        - creation of properties file is done with each change in any of
the controls
                -> this action should only be done once, and the most
convenient place would be when flipping to next page. Unfortunately, the
method that checks if we can flip to the next page is called at each
change on the controls, in so re-creating the file with each keystroke.
The creation of this file was moved on the YoctoBSP Wizard, when
clicking on finish.[v1](update : flipping to properties page fails
while trying to retrieve properties file)
The creation of this file will only be done once at the first
validation of the main page [v2].

        In order to reuse the code that sends a process in the background
and shows the progress from the properties page, we use a
generic class BSPThread that runs a command from the shell in the
background and collects the output & error, and a BSPProgressDialog to
show. Each command will customize the processing of its output by
subclassing BSPThread. the progress of the background thread.

Signed-off-by: Ioana Grigoropol <ioanax.grigoropol@intel.com>