plugins/sdk.ide: Allow using a YoctoUIElement to set the input of a yocto settings widget
Currently a YoctoUIElement is used to create a YoctoUISetting
widget. It is also possible to retrieve a YoctoUIElement containing
the current input of the widget. This patch adds the capability to
also set the input of the widget using a YoctoUIElement. Instead of
using the fControl array clients can now use this method to set the
input of the widget.