plugins/sdk.ide: Move SDK check functionality to separate class
The class YoctoSDKChecker provided some unused code for checking a
Yocto SDK configuration. Checking of the SDK was done in
YoctoSDKUtils. An enum for results was provided in both classes (under
the same name) but contained different possible values.

To avoid confusion, reduce duplicate code and to separate the checking
functionality from the rest of the utility functions, the checking
code has been moved completely to the YoctoSDKChecker.

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>