plugins/sdk.ide: Show SDK check errors in message area
If settings made by the user do not pass the SDK check the error
message was not shown in eclipse but was printed to system out which
was never visible by the user.

The error is now shown in the message area of the property or preference
page if the check fails.

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>