plugins/sdk.ide: Fix setting layout of parent composite
The layout of the parent composite is set in every yocto settings
element. Depending on the order in which the elements are added to the
composite this may lead to different layouts being used.

Layout of the composite is now set in the preference/property pages
instead of setting this in every child component.

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>