plugins/cmake: Allow multiple targets when logging process output
Until now the stream pipe was connecting one input stream with one
output stream. This for example allowed showing the output of a process
in an eclipse console. If the output is required elsewhere, e.g. in a
message dialog, the only possiblity was to read out the whole console
afterwards.

The stream pipe is extended so that we can now write the process output
to multiple target while the process is running, e.g. a eclipse console
and a byte stream.

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>