plugins/cmake: Add cmake error message to error dialog
In case cmake reported an error during project configuration the
resulting error dialog showed a generic 'build failed' message. The
specific error message which can help the user find the problem was
hidden in cmake console.

Along with logging the cmake error to the configure console it is now also
captured and added to the error dialog.

[BUGFIX #6313]

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>