plugins/cmake: Refactor creation of CMAKE_FIND_ROOT_PATH
Adding new environment variables to the list of variables that
contribute to the the CMAKE_FIND_ROOT_PATH involved manually appending
to the value string and also adding a whitespace to separate values.

The construction of the CMAKE_FIND_ROOT_PATH value is extracted to a
separate method, which expects a list of environment variables. Adding
or removing environment variables that contribute to the path is now
achieved by modifying the entries of the list.

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>