plugins/cmake: Add oecore sysroot variables to CMAKE_FIND_ROOT_PATH
The sysroots provided by the toolchain were not added to the
CMAKE_FIND_ROOT_PATH variable. This resolved into CMake searching on the
host system breaking cross-compilation in some cases.

With the addition of the sysroot environment variables provided by
oecore, we make sure that these sysroots are scanned when CMake is
conducting searches.

[BUGFIX #6314]

Signed-off-by: Timo Mueller <timo.mueller@bmw-carit.de>