OptionsPage.java: improved handling of path concatenation in new project
When creating a new Yocto Bitbake Commander project, the folder
separator is not being handled correctly. The problem is a folder and
a file to be concatenated are joined as Strings instead of paths.

This patch solves the issue by using a path joining mechanism instead
of simply concatenating strings.

[Yocto #9648]

Signed-off-by: Humberto Ibarra <humberto.ibarra.lopez@intel.com>
Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>