eclipse-yocto-photonm6.target: add target platform
This staging target platform includes Eclipse Platform 4.8.0 with CDT 9.5.0
and Docker Tooling 4.0.0 contributions for Photon M6 build, this enables
eclipse-yocto to develop against the upcoming release's milestone builds ahead
of the general availability of CDT and Docker Tooling.

The other motivation is to use this separate target platform for tracking the
upcoming release's milestone or RC builds in the event eclipse-yocto.target
still needs to track previous release's maintenance builds.

Signed-off-by: Chin Huat Ang <chin.huat.ang@intel.com>